client/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.27.1
google-generativeai==0.8.3
pydantic==2.6.1
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.28.1
pandas==2.2.0
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.18.0
scikit-learn==1.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
aiofiles==23.2.1
pymongo==4.6.1 